Move history operations to subdirectory
