+++ client/src/components/History/CurrentHistory/HistoryOperations/SelectionStatus.vue
